Made absolute paths to our custom images relative
I search-and-replaced both of these strings with the empty string:

 https://github.com/thinkaurelius/titan/raw/master/doc/
 https://raw.github.com/thinkaurelius/titan/master/doc/

These were all embedded images (the next part of the URL was always
"images/")

The point of this change is to make it possible to export a
self-contained copy of the wiki automatically using gollum-site.  The
wiki uses some external images (e.g. Cassandra logo) that still have
absolute URLs, but the images that live inside the wiki can and
probably should use relative URLs.